filter task tests event
